backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
httpx
python-multipart
psycopg2-binary
aiofiles
motor
